requirements.txt
XlsxWriter==1.3.7
nltk==3.6.5
openpyxl==3.0.5
openpyxl==3.0.5
pandas==1.2.0
python-docx==0.8.11
scipy==1.6.0
urllib3==1.26.7
xlrd==1.2.0
xlwt==1.3.0
spacy
tweepy
requests
newspaper3k
flask
vaderSentiment
validators
